BAC-31 refactor(contract): sales 도메인으로 패키지 이동
