--- src/main/java/com/monsoon/seedflowplus/domain/document/contract/entity/BillingCycle.java
@@ -1,7 +1,0 @@
-package com.monsoon.seedflowplus.domain.document.contract.entity;
-
-public enum BillingCycle {
-    MONTHLY, // 매달
-    QUARTERLY, // 분기
-    HALF_YEARLY // 반기
-}
