--- src/main/java/com/monsoon/seedflowplus/domain/document/order/dto/request/OrderDetailRequest.java
@@ -1,10 +1,0 @@
-package com.monsoon.seedflowplus.domain.document.order.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class OrderDetailRequest {
-
-    private Long contractDetailId;
-    private Long quantity;
-}
