--- src/main/java/com/monsoon/seedflowplus/domain/document/order/dto/response/OrderResponse.java
@@ -1,33 +1,0 @@
-package com.monsoon.seedflowplus.domain.document.order.dto.response;
-
-import com.monsoon.seedflowplus.domain.document.order.entity.OrderStatus;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Getter
-@Builder
-public class OrderResponse {
-
-    private Long orderId;
-    private String orderCode;
-    private Long contractId;
-    private Long clientId;
-    private Long employeeId;
-
-    private BigDecimal totalAmount;
-    private OrderStatus status;
-    private LocalDateTime createdAt;
-
-    // 배송 정보
-    private String shippingName;
-    private String shippingPhone;
-    private String shippingAddress;
-    private String shippingAddressDetail;
-    private String deliveryRequest;
-
-    private List<OrderDetailResponse> items;
-}
