--- src/main/java/com/monsoon/seedflowplus/domain/document/request/entity/QuotationRequestStatus.java
@@ -1,7 +1,0 @@
-package com.monsoon.seedflowplus.domain.document.request.entity;
-
-public enum QuotationRequestStatus {
-    PENDING, // 대기
-    REVIEWING, // 검토
-    COMPLETED // 완료
-}
