+++ src/main/java/com/monsoon/seedflowplus/domain/payment/entity/Payment.java
@@ -2,7 +2,7 @@ package com.monsoon.seedflowplus.domain.payment.entity;
 
 import com.monsoon.seedflowplus.core.common.entity.BaseCreateEntity;
 import com.monsoon.seedflowplus.domain.account.entity.Client;
-import com.monsoon.seedflowplus.domain.document.invoice.entity.Invoice;
+import com.monsoon.seedflowplus.domain.sales.invoice.entity.Invoice;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
