+++ src/main/java/com/monsoon/seedflowplus/domain/sales/contract/entity/BillingCycle.java
@@ -1,0 +1,7 @@
+package com.monsoon.seedflowplus.domain.sales.contract.entity;
+
+public enum BillingCycle {
+    MONTHLY, // 매달
+    QUARTERLY, // 분기
+    HALF_YEARLY // 반기
+}
