+++ src/main/java/com/monsoon/seedflowplus/domain/sales/order/dto/request/OrderCreateRequest.java
@@ -1,0 +1,20 @@
+package com.monsoon.seedflowplus.domain.sales.order.dto.request;
+
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+public class OrderCreateRequest {
+
+    private Long contractId;
+
+    // 배송 정보
+    private String shippingName;
+    private String shippingPhone;
+    private String shippingAddress;
+    private String shippingAddressDetail;
+    private String deliveryRequest;
+
+    private List<OrderDetailRequest> items;
+}
