+++ src/main/java/com/monsoon/seedflowplus/domain/sales/order/dto/request/OrderDetailRequest.java
@@ -1,0 +1,10 @@
+package com.monsoon.seedflowplus.domain.sales.order.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class OrderDetailRequest {
+
+    private Long contractDetailId;
+    private Long quantity;
+}
