+++ src/main/java/com/monsoon/seedflowplus/domain/sales/request/entity/QuotationRequestStatus.java
@@ -1,0 +1,7 @@
+package com.monsoon.seedflowplus.domain.sales.request.entity;
+
+public enum QuotationRequestStatus {
+    PENDING, // 대기
+    REVIEWING, // 검토
+    COMPLETED // 완료
+}
